connect four game
index file is html code file which is an main while style.css is an css file of stylesheet and script.js is an JavaScript file main logic file 

It is the Connect 4 game but web version with a player versus player option, as well as player versus CPU that uses a small artificial intelligence. Throughout the post I will show you how the game works, what technologies I have used, styles, etc. I will also show you how to download the source code, as the game is totally free and open source.